requirements.txt
appnope==0.1.2
backcall==0.2.0
beautifulsoup4==4.9.1
better-profanity==0.7.0
boto3==1.8.9
botocore==1.11.9
certifi==2020.4.5.2
chardet==3.0.4
decorator==5.0.7
Django==2.1.1
django-el-pagination==3.2.4
django-htmlmin==0.10.0
django-jet==1.0.7
django-markdown-deux==1.0.5
django-multiselectfield==0.1.8
django-pagedown==1.0.6
django-rest-framework==0.1.0
django-runas==0.1.3
django-s3direct==1.0.4
djangorestframework==3.11.0
docutils==0.16
html5lib==1.0.1
idna==2.7
ipdb==0.13.7
ipython==7.23.1
ipython-genutils==0.2.0
jedi==0.18.0
jmespath==0.10.0
markdown2==2.3.9
matplotlib-inline==0.1.2
parso==0.8.2
pexpect==4.8.0
phonenumbers==8.12.5
pickleshare==0.7.5
prompt-toolkit==3.0.18
psycopg2==2.7.5
ptyprocess==0.7.0
Pygments==2.9.0
PyJWT==1.7.1
python-dateutil==2.8.1
pytz==2020.1
requests==2.19.1
s3transfer==0.1.13
sendwithus==5.0.0
sentry-sdk==0.14.4
short-url==1.2.2
six==1.15.0
soupsieve==2.0.1
toml==0.10.2
traitlets==5.0.5
treelib==1.5.3
twilio==6.42.0
urllib3==1.23
wcwidth==0.2.5
webencodings==0.5.1
